[IEE-57] Cleanup (#34)
Simplified DI
Simplified UseCases
Created Convention plugin for creating flavors
Configured flavors for local and production servers
Removed http logging from production builds
Used injections for dispatchers instead of hardcoding them
Wrapped refreshes with Result<> to observe success and errors in upper layers
Created crashlytics module for logging non fatal exceptions
Simplified State management in certain screens
Added toast notifications on certain errors
Created work module for scheduling and executing background work
Created worker for notifying users about new announcements
Removed FCM topics support

+++ core/testing/src/main/java/com/kastik/apps/core/testing/datasource/local/FakeAuthenticationLocalDataSource.kt
@@ -17,8 +17,8 @@ class FakeAuthenticationLocalDataSource : AuthenticationLocalDataSource {
     private val _aboardTokenExpiration = MutableStateFlow<Int?>(null)
     val aboardTokenExpiration: StateFlow<Int?> = _aboardTokenExpiration.asStateFlow()
 
-    private val _aboardTokenLastRefreshTime = MutableStateFlow<Int?>(null)
-    val aboardTokenLastRefreshTime: StateFlow<Int?> = _aboardTokenLastRefreshTime.asStateFlow()
+    private val _aboardTokenLastRefreshTime = MutableStateFlow<Long?>(null)
+    val aboardTokenLastRefreshTime: StateFlow<Long?> = _aboardTokenLastRefreshTime.asStateFlow()
 
     override fun getIsSignedIn(): Flow<Boolean> {
         return isSignedIn
@@ -34,7 +34,7 @@ class FakeAuthenticationLocalDataSource : AuthenticationLocalDataSource {
         _aboardTokenExpiration.value = expiration
     }
 
-    override suspend fun setAboardTokenLastRefreshTime(lastRefreshTime: Int) {
+    override suspend fun setAboardTokenLastRefreshTime(lastRefreshTime: Long) {
         _aboardTokenLastRefreshTime.value = lastRefreshTime
     }
 
@@ -42,7 +42,7 @@ class FakeAuthenticationLocalDataSource : AuthenticationLocalDataSource {
         return aboardTokenExpiration
     }
 
-    override fun getAboardTokenLastRefreshTime(): Flow<Int?> {
+    override fun getAboardTokenLastRefreshTime(): Flow<Long?> {
         return aboardTokenLastRefreshTime
     }
